\name{Rassoc-package}
\Rdversion{1.1}
\alias{Rassoc-package}
\alias{Rassoc}
\docType{package}
\title{
Robust tests for case-control genetic association studies 
}
\description{
This package supplies several robust tests for case-control genetic association studies. The tests contained in this package are: allelic based test, Cochran-Armitage trend test, maximin efficiency robust test, MAX3 test and genetic model selection test. For each test, the corresponding R code reports the test statistics, the associated p-value and conclusion of hypothesis test under significant level 0.05 respectively. 
}
\seealso{\code{\link{ABT}},\code{\link{CATT}},\code{\link{GMS}},\code{\link{MAX3}},\code{\link{MERT}}}
\details{
\tabular{ll}{
Package: \tab Rassoc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-07-21\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Yong ZANG <zanngyong@hku.hk>,
Wingkam FUNG <wingfung@hku.hk>,
Gang ZHENG <zhengg@nhlbi.nih.gov> 

Maintainer: Yong ZANG <zanngyong@hku.hk>
}
\references{
Zang Y, Fung WK and Zheng G(2009). Simple Algorithms to calculate the asymptotic null distribution for MAX3 and genetic model selection in case-control genetic association studies in R. Journal of Statistical software (revised).  
}

