\name{Rassoc-package}
\Rdversion{1.1}
\alias{Rassoc-package}
\alias{Rassoc}
\docType{package}
\title{
Robust tests for case-control genetic association studies 
}
\description{
This package supplies several robust tests for case-control genetic association studies. The tests contained in this package are: allelic based test, Cochran-Armitage trend test, maximin efficiency robust test, MAX3 test and genetic model selection test. For each test, the corresponding R code reports the test statistics and the associated p-value. 
}
\seealso{\code{\link{ABT}},\code{\link{CATT}},\code{\link{GMS}},\code{\link{MAX3}},\code{\link{MERT}}}
\details{
\tabular{ll}{
Package: \tab Rassoc\cr
Type: \tab Package\cr
Version: \tab 1.01\cr
Date: \tab 2009-11-06\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Yong ZANG <zanngyong@hku.hk>,
Wingkam FUNG <wingfung@hku.hk>,
Gang ZHENG <zhengg@nhlbi.nih.gov> 

Maintainer: Yong ZANG <zanngyong@hku.hk>
}
\references{
Zang Y, Fung WK and Zheng G(2010). Simple algorithms to calculate the asymptotic null distributions of robust tests in case-control genetic association studies in R. Journal of Statistical software 33(8).  
}

